refactor forbidden components
